+ Redesign tbl_Prices + Added SPs to calc MA50,MA200 and EMA25 + deleted old SPs that were calculating the avreages

+++ sql/DataContract.sql
@@ -85,7 +85,7 @@ while exists(select * from @tbl where idx = @i)
 	set @tempMIDDLE = REPLACE(@tempMIDDLE, 'real', 'Single')
 	set @tempMIDDLE = REPLACE(@tempMIDDLE, 'decimal', 'double')
 	set @tempMIDDLE = REPLACE(@tempMIDDLE, 'tinyint', 'Byte')
-	set @tempMIDDLE = REPLACE(@tempMIDDLE, 'money', 'double')
+	set @tempMIDDLE = REPLACE(@tempMIDDLE, 'money', 'decimal')
 	set @tempMIDDLE = REPLACE(@tempMIDDLE, 'numeric', 'long')
